✨ Improved leaderboard paged performance
